# rest/server/delete_visita.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/gofrs/uuid"

	"taqueria/core/module/visita/types"
)

func (h *handlers) DeleteVisita(w http.ResponseWriter, r *http.Request) {

	IDStr := chi.URLParam(r, "id")
	IDVal, err := uuid.FromString(IDStr)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid path parameter ID: %w", err))
		return
	}

	if delErr := h.core.Visita().Delete(r.Context(), types.DeleteRequest{
		ID: IDVal,
	}); delErr != nil {
		if delErr.Error() == "entity not found" {
			writeProblem(w, http.StatusNotFound, delErr)
			return
		}
		writeProblem(w, http.StatusInternalServerError, delErr)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}