# rest/server/get_taco.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/gofrs/uuid"

	"taqueria/core/module/taco/types"
)

func (h *handlers) GetTaco(w http.ResponseWriter, r *http.Request) {

	IDStr := chi.URLParam(r, "id")
	IDVal, err := uuid.FromString(IDStr)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid path parameter ID: %w", err))
		return
	}

	fetchRes, err := h.core.Taco().FetchTacoByID(r.Context(), types.FetchTacoByIDRequest{
		ID: IDVal,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusNotFound, fmt.Errorf("entity not found"))
		return
	}

	writeJSON(w, http.StatusOK, fetchRes.Results[0])
}