# rest/server/list_taco.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	tacomodule "taqueria/core/module/taco"
	"taqueria/core/module/taco/types"
	main_entity "taqueria/entity/taco"
)

func (h *handlers) ListTaco(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := tacoDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            tacoEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.Taco().List(r.Context(), req, tacomodule.WithSkipCache())
	} else {
		result, err = h.core.Taco().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.Taco]{
		Items: result.Taco,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func tacoDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// taco

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("taqueria_uuid", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in tacoEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("tipo", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_INVALID", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_PASTOR", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_SUADERO", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_CAMPECHANO", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_BISTEC", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_LENGUA", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_TRIPA", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_CHORIZO", filtering.TypeString),
		filtering.DeclareIdent("TIPO_TACO_CANASTA", filtering.TypeString),

		filtering.DeclareIdent("precio", filtering.TypeFloat),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for taco: %v", err)
	}
	return declarations
}

// tacoEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func tacoEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"tipo": {
			"invalid":              0,
			"TIPO_TACO_INVALID":    0,
			"pastor":               1,
			"TIPO_TACO_PASTOR":     1,
			"suadero":              2,
			"TIPO_TACO_SUADERO":    2,
			"campechano":           3,
			"TIPO_TACO_CAMPECHANO": 3,
			"bistec":               4,
			"TIPO_TACO_BISTEC":     4,
			"lengua":               5,
			"TIPO_TACO_LENGUA":     5,
			"tripa":                6,
			"TIPO_TACO_TRIPA":      6,
			"chorizo":              7,
			"TIPO_TACO_CHORIZO":    7,
			"canasta":              8,
			"TIPO_TACO_CANASTA":    8,
		},
	}
}